requirements.txt
flask
flask_sqlalchemy
flask_login
python-dotenv
huggingface_hub
reportlab
requests
beautifulsoup4
pypdf
youtube-transcript-api
openai
